COMMERCE-11740 autogenerated code (gw buildService)

+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/model/CommerceInventoryAuditModel.java
@@ -20,6 +20,8 @@ import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.model.MVCCModel;
 import com.liferay.portal.kernel.model.ShardedModel;
 
+import java.math.BigDecimal;
+
 import java.util.Date;
 
 import org.osgi.annotation.versioning.ProviderType;
@@ -222,14 +224,14 @@ public interface CommerceInventoryAuditModel
 	 *
 	 * @return the quantity of this commerce inventory audit
 	 */
-	public int getQuantity();
+	public BigDecimal getQuantity();
 
 	/**
 	 * Sets the quantity of this commerce inventory audit.
 	 *
 	 * @param quantity the quantity of this commerce inventory audit
 	 */
-	public void setQuantity(int quantity);
+	public void setQuantity(BigDecimal quantity);
 
 	/**
 	 * Returns the sku of this commerce inventory audit.
